requirements.txt
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.8.30
chardet==3.0.4
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
deep-translator==1.11.4
deepl==1.20.0
google-api-core==2.21.0
google-auth==2.35.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-cloud-translate==3.16.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
googletrans==4.0.0rc1
grpc-google-iam-v1==0.13.1
grpcio==1.67.0
grpcio-status==1.67.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.10.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
libretranslatepy==2.1.1
lxml==5.3.0
numpy==2.1.3
pandas==2.2.3
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
translate==3.6.1
tzdata==2024.2
urllib3==2.2.3
